+++ docs/changelog.md
@@ -61,6 +61,8 @@ Changed:
 - [imshow2()][extra.utils.imshow2] will now add a colorbar automatically (!351).
 - [Calibration](calibration.md) conditions objects can now be displayed as
   Markdown in Jupyter notebooks (!381).
+- Increased maximal length of pulse masks returned by
+  [pulse_mask()][extra.components.pulses.PulsePattern.pulse_mask] to 4096 (!461).
 - [CookieboxCalibration][extra.applications.CookieboxCalibration] calculates the transmission
   using the sum of the trace instead of the fit amplitude for robustness.